BLOCK-277:decodeBaseAccount method created
